fix(highs): pass uint8 integrality array to changeColsIntegrality
highspy's changeColsIntegrality expects var_types as a uint8 array
(matching HiGHS's uint8-backed HighsVarType enum). The persistent
update path already builds it as uint8 (_apply_var_types); align the
direct-solve path so it matches highspy's tightened type stubs and is
robust across versions.

+++ linopy/solvers.py
@@ -1656,7 +1656,7 @@ class Highs(Solver[None]):
             int_mask = (vtypes == "B") | (vtypes == "I") | (vtypes == "S")
             labels = np.arange(len(vtypes))[int_mask]
             integrality = np.array(
-                [integrality_map[v] for v in vtypes[int_mask]], dtype=np.int32
+                [integrality_map[v] for v in vtypes[int_mask]], dtype=np.uint8
             )
             h.changeColsIntegrality(len(labels), labels, integrality)
